services/rules-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
jsonschema==4.20.0
python-dotenv==1.0.0
structlog==23.2.0

